DEV: Update agent core source files for v3.0.0
Update Ioto agent core for v3.0.0 features:

- SERVICES_REGISTER support separate from SERVICES_CLOUD
- cmdBuilder configuration available for provision-only builds
- Improved error handling and diagnostics
- Update command-line tools (ioto, json, url, web)
- Rename url --count to --iterations

+++ src/cloud/sync.c
@@ -631,7 +631,7 @@ static void cleanSyncChanges(Json *json)
             }
         }
     }
-    rDebug("sync", "After syncing %ld changes %d changes pending", count, rGetHashLength(ioto->syncHash));
+    rDebug("sync", "After syncing %d changes %d changes pending", (int) count, rGetHashLength(ioto->syncHash));
     if (count && rGetHashLength(ioto->syncHash) == 0) {
         //  OPTIMIZE
         recreateSyncLog();
